requirements.txt
flask
python-dotenv
pandas
numpy
scikit-learn
scipy
joblib
tensorflow
sqlalchemy
pymysql